requirements.txt
pandas
requests
dotenv
pyyaml
psycopg2-binary 
sqlalchemy
click